python/requirements.txt
# Google APIs
google-ads>=24.0.0
google-analytics-data>=0.18.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-cloud-bigquery>=3.13.0

# Facebook / Instagram
facebook-business>=19.0.0

# HTTP
requests>=2.31.0

# Config
python-dotenv>=1.0.0
